Conflicts:

app/src/main/java/com/metis/meishuquan/model/provider/ApiDataProvider.ja
va

+++ app/src/main/java/com/metis/meishuquan/model/BLL/CircleOperator.java
@@ -156,7 +156,7 @@ public class CircleOperator {
         }
     }
 
-    public void cancelAttention(int userId, ApiOperationCallback<ReturnInfo<String>> callback) {
+    public void cancelAttention(long userId, ApiOperationCallback<ReturnInfo<String>> callback) {
         if (SystemUtil.isNetworkAvailable(MainApplication.UIContext)) {
             if (flag) {
                 StringBuffer PATH = new StringBuffer(URL_CANCEL_ATTENTION);
